requirements.txt
# Core Dependencies
google-generativeai==0.3.1
langchain-community==0.0.10
langchain-google-genai==0.0.5
chromadb==0.4.15
python-dotenv==1.0.0
typer==0.9.0
pathlib==1.0.1

# Utility Dependencies
rich==13.7.0  # For enhanced terminal output
tqdm==4.66.1  # For progress bars
colorama==0.4.6  # For colored terminal text

# Optional Dependencies
pytest==7.4.3  # For testing
pytest-cov==4.1.0  # For test coverage
black==23.12.1  # For code formatting
flake8==7.0.0  # For linting
mypy==1.8.0  # For type checking